+++ dapiwrap/__init__.py
@@ -1,8 +1,3 @@
 from dapiwrap import DAPIWrap
 from dapiwconst import *
-from dapiwtools import (
-    Downwad, 
-    print_wad_info,
-    random_wad,
-    SearchFilter
-)
+from dapiwtools import *