+++ dapiwrap/dapiwrap.py
@@ -1,27 +1,33 @@
 #===============================================================================
 # DAPIWrap: Doomworld API Wrapper
 #-------------------------------------------------------------------------------
-# Version: 0.1.0
-# Updated: 07-06-2014
+# Version: 0.2.0
+# Updated: 11-06-2014
 # Author: Alex Crawford
 # License: MIT
 #===============================================================================
 
 """
-A Python wrapper for the Doomworld /idgames archive API.
+DAPIWrap, short for "Doomworld API Wrapper", is a simple Python wrapper for the 
+Doomworld idgames archive API.
 
-API documentation:
+This pacakge is still in the beta stages, so expect changes that may break
+things, if you're using it.
+
+Doomworld API documentation:
 http://www.doomworld.com/idgames/api/
 
-Doomworld /idgames archive:
+Doomworld idgames archive:
 http://www.doomworld.com/idgames/
 
-Please, if you are using this, don't hammer the Doomworld API with requests.
-I am not sure what kind of traffic their API can handle, so do the right 
-thing, and limit the number of requests you make. I don't want this wrapper to 
-be used (abused) to bog down their server. Be cool.
+Note:
 
-This script requires the `requests` package:
+Please don't hammer the Doomworld API with requests. I am not sure what kind 
+of traffic their API can handle, so do the right thing, and limit the number 
+of requests you make. I don't want this wrapper to be used (abused) to bog down 
+their server. Be cool.
+
+This package requires the ``requests`` package:
 http://docs.python-requests.org/en/latest/
 
 """
@@ -30,15 +36,8 @@ http://docs.python-requests.org/en/latest/
 # Imports
 #===============================================================================
 
-# from collections import deque
-import random
 import requests
 
-from dapiwtools import (
-    determine_lvl_path,
-    Downwad,
-    SearchFilter
-)
 from dapiwconst import (
     A_ABOUT,
     A_DBPING,
@@ -62,26 +61,19 @@ from dapiwconst import (
     A_SEARCH_TYPE,
     A_OUT_JSON,
     API_URL,
-    DIRECT_ASC,
-    DIRECT_DESC,
-    FILTER_DATE,
-    FILTER_GAME,
-    FILTER_RATING,
-    FILTER_SIZE,
-    FILTER_VOTES,
-    FILTER_YEAR,
-    SORT_DATE,
-    SORT_FILE,
-    SORT_RATING,
-    SORT_SIZE,
     TYPE_AUTHOR,
     TYPE_CREDITS,
-    TYPE_DECRIP,
+    TYPE_DESCRIP,
     TYPE_EDITORS,
     TYPE_EMAIL,
     TYPE_FILE,
     TYPE_TEXT,
     TYPE_TITLE,
+)
+
+from dapiwtools import (
+    Downwad,
+    SearchFilter
 )
 
 #===============================================================================
@@ -90,8 +82,6 @@ from dapiwconst import (
 
 class DAPIWrap(object):
     """The main DAPIWrap class."""
-
-    # CACHE_LIMIT = 100
 
     A_NOPARAM = [
         A_ABOUT, A_DBPING, A_PING
@@ -110,10 +100,9 @@ class DAPIWrap(object):
         :param dl_folder: The location to download files to.
 
         """
-        # self.search_cache = deque([])
-        # self.wadinfo_cache = deque([])
         self.dl_folder = dl_folder
 
+        # DAPIWrap tools
         self.download = Downwad(self)
         self.filter = SearchFilter()
 
@@ -126,7 +115,7 @@ class DAPIWrap(object):
         """
         about = self.call(A_ABOUT)
 
-        if not raw and about.get("content"):
+        if not raw and "content" in about:
             return about["content"]
         else:
             return about
@@ -153,11 +142,11 @@ class DAPIWrap(object):
             if action == A_SEARCH:
                 if type(params) == dict:
                     url = "%s%s" % (API_URL, A_SEARCH_QUERY % (params["query"]))
-                    if params.get("type"):
+                    if "type" in params:
                         url += A_SEARCH_TYPE % (params["type"])
-                    if params.get("sort"):
+                    if "sort" in params:
                         url += A_SEARCH_SORT % (params["sort"])
-                    if params.get("dir"):
+                    if "dir" in params:
                         url += A_SEARCH_DIRECT % (params["dir"])
                 else:
                     url = "%s%s" % (API_URL, A_SEARCH_QUERY % (params))
@@ -166,17 +155,6 @@ class DAPIWrap(object):
         data = requests.get(url).json()
 
         return data
-
-    # def cache_add(self, cache, items):
-    #     """Adds the given items to the given cache."""
-
-    #     if type(items) == list:
-    #         [cache.append(x) for x in items if x not in cache]
-    #     else:
-    #         cache.append(items)
-
-    #     while len(cache) >= self.CACHE_LIMIT:
-    #         cache.popleft()
 
     def dbping(self):
         """
@@ -191,7 +169,7 @@ class DAPIWrap(object):
         """
         Gets the contents of a given path.
 
-        :param path: The /idgames path to get content from.
+        :param path: The idgames path to get content from.
         :param raw: Whether to return the data exactly as recieved (raw), or
         to extract the contents and return just the contents in a list.
 
@@ -200,12 +178,12 @@ class DAPIWrap(object):
         """
         contents = self.call(A_GETCONTENTS_NAME, path)
 
-        if not raw and contents.get("content"):
+        if not raw and "content" in contents:
             return contents["content"]
         else:
             return contents
 
-    def get_file(self, path, raw=False):
+    def get_file_path(self, path, raw=False):
         """
         Gets a file from the given path.
 
@@ -217,32 +195,8 @@ class DAPIWrap(object):
 
         """
         wad_info = self.call(A_GET_FILE, path)
-        # self.cache_add(self.wadinfo_cache, wad_info)
 
-        if not raw and wad_info.get("content"):
-            # self.cache_add(self.wadinfo_cache, wad_info["content"])
-            return wad_info["content"]
-        else:
-            return wad_info
-
-    def get_file_alt(self, filename, game, raw=False):
-        """
-        Gets a file using the given filename.
-
-        :param filename: The filename of the wad.
-        :param game: The game the wad is for.
-        :param raw: Whether to return the data exactly as recieved (raw), or
-        to extract the contents and return just the contents in a list.
-
-        :returns: Wad info for the filename.
-
-        """
-        path = "%s%s" % (determine_lvl_path(filename, game), filename)
-
-        wad_info = self.call(A_GET_FILE, path)
-
-        if not raw and wad_info.get("content"):
-            # self.cache_add(self.wadinfo_cache, wad_info["content"])
+        if not raw and "content" in wad_info:
             return wad_info["content"]
         else:
             return wad_info
@@ -251,7 +205,7 @@ class DAPIWrap(object):
         """
         Gets the files under a directory of the given path.
 
-        :param path: The /idgames path to return files from.
+        :param path: The idgames path to return files from.
         :param raw: Whether to return the data exactly as recieved (raw), or
         to extract the contents and return just the contents in a list.
 
@@ -260,17 +214,19 @@ class DAPIWrap(object):
         """
         files = self.call(A_GETFILES_NAME, path)
 
-        if not raw and files.get("content"):
-            # self.cache_add(self.wadinfo_cache, files["content"]["file"])
-            return files["content"]["file"]
+        if not raw and "content" in files:
+            if type(files["content"]["file"]) == list:
+                return files["content"]["file"]
+            else:
+                return [files["content"]["file"]]
         else:
-            return files
+            return []
 
     def get_dirs(self, path, raw=False):
         """
         Gets the directories under a directory of the given path.
 
-        :param path: The /idgames path to return directories from.
+        :param path: The idgames path to return directories from.
         :param raw: Whether to return the data exactly as recieved (raw), or
         to extract the contents and return just the contents in a list.
 
@@ -279,7 +235,7 @@ class DAPIWrap(object):
         """
         dirs = self.call(A_GETDIRS_NAME, path)
 
-        if not raw and dirs.get("content"):
+        if not raw and "content" in dirs:
             return dirs["content"]
         else:
             return dirs
@@ -296,13 +252,33 @@ class DAPIWrap(object):
 
         """
         wad_info = self.call(A_GET_ID, wad_id)
-        # self.cache_add(self.wadinfo_cache, wad_info)
 
-        if not raw and wad_info.get("content"):
-            # self.cache_add(self.wadinfo_cache, wad_info["content"])
+        if not raw and "content" in wad_info:
             return wad_info["content"]
         else:
             return wad_info
+
+    def get_id_list(self, id_list, raw=False):
+        """
+        Gets the info for a wad with the given ID.
+
+        :param id_list: A list of wad ID numbers.
+        :param raw: Whether to return the data exactly as recieved (raw), or
+        to extract the contents and return just the contents in a list.
+
+        :returns: Wad info for the given list of IDs.
+
+        """
+        results = []
+
+        for wad_id in id_list:
+            wad_info = self.call(A_GET_ID, wad_id)
+            if not raw and "content" in wad_info:
+                results.append(wad_info["content"])
+            else:
+                return wad_info
+
+        return results
 
     def get_latestfiles(self, limit=10):
         """
@@ -315,8 +291,7 @@ class DAPIWrap(object):
         """
         latest = self.call(A_LATESTFILES, limit)
 
-        if not raw and latest.get("content"):
-            # self.cache_add(self.wadinfo_cache, latest["content"])
+        if not raw and "content" in latest:
             return latest["content"]
         else:
             return latest
@@ -332,8 +307,7 @@ class DAPIWrap(object):
         """
         latest = self.call(A_LATESTVOTES, limit)
 
-        if not raw and latest.get("content"):
-            # self.cache_add(self.wadinfo_cache, latest["content"])
+        if not raw and "content" in latest:
             return latest["content"]
         else:
             return latest
@@ -342,7 +316,7 @@ class DAPIWrap(object):
         """
         Gets the parent directory for a given path.
 
-        :param path: The /idgames path to return the parent from.
+        :param path: The idgames path to return the parent from.
         :param raw: Whether to return the data exactly as recieved (raw), or
         to extract the contents and return just the contents in a list.
 
@@ -351,7 +325,7 @@ class DAPIWrap(object):
         """
         parent_dir = self.call(A_PARENTDIR_FILE, path)
 
-        if not raw and parent_dir.get("content"):
+        if not raw and "content" in parent_dir:
             return parent_dir["content"]
         else:
             return parent_dir
@@ -367,7 +341,7 @@ class DAPIWrap(object):
 
     def search(self, query, params={}):
         """
-        Search the /idgames archive.
+        Search the idgames archive.
 
         :param query: The search query.
         :param params: A ``dict`` of search parameters and/or filters.
@@ -396,19 +370,19 @@ class DAPIWrap(object):
             Accepted Values:
                 FILTER_GAME, FILTER_GYR, FILTER_YEAR, FILTER_RATING
 
-        :returns: The search results.
+        :returns: A list of search results.
 
         """
         params["query"] = query
 
         results = self.call(A_SEARCH, params)
 
-        if not params.get("raw") and results.get("content"):
+        if "raw" not in params and "content" in results:
             if type(results["content"]["file"]) == list:
                 results = results["content"]["file"]
             else:
                 results = [results["content"]["file"]]
-            if params.get("filter"):
+            if "filter" in params:
                 if type(params["filter"][0]) not in (tuple, list):
                     srchfilter = params["filter"][0]
                     filterargs = params["filter"][1]
@@ -422,13 +396,12 @@ class DAPIWrap(object):
                         results = self.filter.filters[srchfilter](
                             results, filterargs
                         )
-            # self.cache_add(self.search_cache, results)
-
-        return results
+            return results
+        return []
 
     def search_author(self, query, params={}):
         """
-        Search the /idgames archive for the given author, using the 
+        Search the idgames archive for the given author, using the 
         default search parameters.
 
         :param query: The search query.
@@ -445,7 +418,7 @@ class DAPIWrap(object):
 
     def search_credits(self, query, params={}):
         """
-        Search the credits of the files in the /idgames archive, using the 
+        Search the credits of the files in the idgames archive, using the 
         default search parameters.
 
         :param query: The search query.
@@ -462,7 +435,7 @@ class DAPIWrap(object):
 
     def search_descrip(self, query, params={}):
         """
-        Search the descriptions of files in the /idgames archive, using 
+        Search the descriptions of files in the idgames archive, using 
         the default search parameters.
 
         :param query: The search query.
@@ -479,7 +452,7 @@ class DAPIWrap(object):
 
     def search_editors(self, query, params={}):
         """
-        Search the /idgames archive for the given editors, using the 
+        Search the idgames archive for the given editors, using the 
         default search parameters.
 
         :param query: The search query.
@@ -496,7 +469,7 @@ class DAPIWrap(object):
 
     def search_email(self, query, params={}):
         """
-        Search the /idgames archive for the given email, using the 
+        Search the idgames archive for the given email, using the 
         default search parameters.
 
         :param query: The search query.
@@ -513,7 +486,7 @@ class DAPIWrap(object):
 
     def search_text(self, query, params={}):
         """
-        Search the /idgames archive for the given title, using the 
+        Search the idgames archive for the given title, using the 
         default search parameters.
 
         :param query: The search query.
@@ -530,7 +503,7 @@ class DAPIWrap(object):
 
     def search_title(self, query, params={}):
         """
-        Search the /idgames archive for the given title, using the 
+        Search the idgames archive for the given title, using the 
         default search parameters.
 
         :param query: The search query.